+++ src/main/java/es/in2/issuer/backend/shared/domain/model/dto/AccessTokenContext.java
@@ -1,5 +1,8 @@
 package es.in2.issuer.backend.shared.domain.model.dto;
 
+import lombok.Builder;
+
+@Builder
 public record AccessTokenContext(
         String rawToken,
         String jti,