+++ src/main/java/es/in2/issuer/backend/shared/domain/service/RefreshTokenService.java
@@ -1,0 +1,8 @@
+package es.in2.issuer.backend.shared.domain.service;
+
+import java.time.Instant;
+
+public interface RefreshTokenService {
+    String generateRefreshToken();
+    long generateRefreshTokenExpirationTime(Instant issueTime);
+}